Allow to explicitly set debug mode
Allow to explicitly set debug mode instead of relying on the logger. Running `--debug` in pelican only changes the logging so it can be used in production however with this using the logger to change how assets work I can no longer use it in production (which is needed for other reasons). So this change is to allow more control.

I haven't tested or anything, about to make the change myself and use it. However adding code here should easily show you what  I mean if you want to implement it a different way.